Introduce `tun` crate
