+++ rust/bin-shared/src/tun_device_manager/linux.rs
@@ -1,19 +1,38 @@
 //! Virtual network interface
 
+use crate::FIREZONE_MARK;
 use anyhow::{anyhow, Context as _, Result};
 use connlib_shared::DEFAULT_MTU;
-use firezone_tunnel::Tun;
 use futures::TryStreamExt;
 use ip_network::{IpNetwork, Ipv4Network, Ipv6Network};
+use libc::{close, fcntl, makedev, mknod, open, F_GETFL, F_SETFL, O_NONBLOCK, O_RDWR, S_IFCHR};
 use netlink_packet_route::route::{RouteProtocol, RouteScope};
 use netlink_packet_route::rule::RuleAction;
 use rtnetlink::{new_connection, Error::NetlinkError, Handle, RouteAddRequest, RuleAddRequest};
+use std::path::Path;
+use std::task::{Context, Poll};
 use std::{
     collections::HashSet,
     net::{Ipv4Addr, Ipv6Addr},
 };
+use std::{
+    ffi::CStr,
+    fs, io,
+    os::{
+        fd::{AsRawFd, RawFd},
+        unix::fs::PermissionsExt,
+    },
+};
+use tokio::io::unix::AsyncFd;
+use tun::ioctl;
+use tun::TunTrait;
 
-use crate::FIREZONE_MARK;
+const TUNSETIFF: libc::c_ulong = 0x4004_54ca;
+const TUN_DEV_MAJOR: u32 = 10;
+const TUN_DEV_MINOR: u32 = 200;
+
+// Safety: We know that this is a valid C string.
+const TUN_FILE: &CStr = unsafe { CStr::from_bytes_with_nul_unchecked(b"/dev/net/tun\0") };
 
 const FILE_ALREADY_EXISTS: i32 = -17;
 const FIREZONE_TABLE: u32 = 0x2021_fd00;
@@ -36,7 +55,7 @@ impl Drop for TunDeviceManager {
 }
 
 impl TunDeviceManager {
-    pub const IFACE_NAME: &'static str = "tun-firezone"; // Keep this synced with `Tun` until we fix the module dependencies (i.e. move `Tun` out of `firezone-tunnel`).
+    pub const IFACE_NAME: &'static str = "tun-firezone";
 
     /// Creates a new managed tunnel device.
     ///
@@ -256,4 +275,126 @@ async fn remove_route(route: &IpNetwork, idx: u32, handle: &Handle) -> Result<()
         .await
         .context("Failed to delete route")?;
     Ok(())
+}
+
+#[derive(Debug)]
+pub struct Tun {
+    fd: AsyncFd<RawFd>,
+}
+
+impl Tun {
+    pub fn new() -> io::Result<Self> {
+        create_tun_device()?;
+
+        let fd = match unsafe { open(TUN_FILE.as_ptr() as _, O_RDWR) } {
+            -1 => return Err(get_last_error()),
+            fd => fd,
+        };
+
+        // Safety: We just opened the file descriptor.
+        unsafe {
+            ioctl::exec(
+                fd,
+                TUNSETIFF,
+                &mut ioctl::Request::<ioctl::SetTunFlagsPayload>::new(TunDeviceManager::IFACE_NAME),
+            )?;
+        }
+
+        set_non_blocking(fd)?;
+
+        // Safety: We just opened the fd.
+        unsafe { Self::from_fd(fd) }
+    }
+
+    /// Create a new [`Tun`] from a raw file descriptor.
+    ///
+    /// # Safety
+    ///
+    /// The file descriptor must be open.
+    unsafe fn from_fd(fd: RawFd) -> io::Result<Self> {
+        Ok(Tun {
+            fd: AsyncFd::new(fd)?,
+        })
+    }
+}
+
+impl Drop for Tun {
+    fn drop(&mut self) {
+        unsafe { close(self.fd.as_raw_fd()) };
+    }
+}
+
+impl TunTrait for Tun {
+    fn write4(&self, buf: &[u8]) -> io::Result<usize> {
+        write(self.fd.as_raw_fd(), buf)
+    }
+
+    fn write6(&self, buf: &[u8]) -> io::Result<usize> {
+        write(self.fd.as_raw_fd(), buf)
+    }
+
+    fn poll_read(&mut self, buf: &mut [u8], cx: &mut Context<'_>) -> Poll<io::Result<usize>> {
+        tun::unix::poll_raw_fd(&self.fd, |fd| read(fd, buf), cx)
+    }
+
+    fn name(&self) -> &str {
+        TunDeviceManager::IFACE_NAME
+    }
+}
+
+fn get_last_error() -> io::Error {
+    io::Error::last_os_error()
+}
+
+fn set_non_blocking(fd: RawFd) -> io::Result<()> {
+    match unsafe { fcntl(fd, F_GETFL) } {
+        -1 => Err(get_last_error()),
+        flags => match unsafe { fcntl(fd, F_SETFL, flags | O_NONBLOCK) } {
+            -1 => Err(get_last_error()),
+            _ => Ok(()),
+        },
+    }
+}
+
+fn create_tun_device() -> io::Result<()> {
+    let path = Path::new(TUN_FILE.to_str().expect("path is valid utf-8"));
+
+    if path.exists() {
+        return Ok(());
+    }
+
+    let parent_dir = path.parent().unwrap();
+    fs::create_dir_all(parent_dir)?;
+    let permissions = fs::Permissions::from_mode(0o751);
+    fs::set_permissions(parent_dir, permissions)?;
+    if unsafe {
+        mknod(
+            TUN_FILE.as_ptr() as _,
+            S_IFCHR,
+            makedev(TUN_DEV_MAJOR, TUN_DEV_MINOR),
+        )
+    } != 0
+    {
+        return Err(get_last_error());
+    }
+
+    Ok(())
+}
+
+/// Read from the given file descriptor in the buffer.
+fn read(fd: RawFd, dst: &mut [u8]) -> io::Result<usize> {
+    // Safety: Within this module, the file descriptor is always valid.
+    match unsafe { libc::read(fd, dst.as_mut_ptr() as _, dst.len()) } {
+        -1 => Err(io::Error::last_os_error()),
+        n => Ok(n as usize),
+    }
+}
+
+/// Write the buffer to the given file descriptor.
+fn write(fd: RawFd, buf: &[u8]) -> io::Result<usize> {
+    // Safety: Within this module, the file descriptor is always valid.
+    match unsafe { libc::write(fd, buf.as_ptr() as _, buf.len() as _) } {
+        -1 => Err(io::Error::last_os_error()),
+        n => Ok(n as usize),
+    }
 }
