+++ rust/connlib/clients/android/Cargo.toml
@@ -23,12 +23,14 @@ phoenix-channel = { workspace = true }
 secrecy = { workspace = true }
 serde_json = "1"
 socket-factory = { workspace = true }
+tun = { workspace = true }
 thiserror = "1"
 tokio = { workspace = true, features = ["rt"] }
 tracing = { workspace = true, features = ["std", "attributes"] }
 tracing-appender = "0.2"
 tracing-subscriber = { workspace = true }
 url = "2.4.0"
+libc = "0.2"
 
 [target.'cfg(target_os = "android")'.dependencies]
 android_log-sys = "0.3.1"
