+++ rust/connlib/clients/apple/src/lib.rs
@@ -2,11 +2,12 @@
 #![allow(clippy::unnecessary_cast, improper_ctypes, non_camel_case_types)]
 
 mod make_writer;
+mod tun;
 
 use backoff::ExponentialBackoffBuilder;
 use connlib_client_shared::{
     callbacks::ResourceDescription, file_logger, keypair, Callbacks, ConnectArgs, Error, LoginUrl,
-    Session, Tun, V4RouteList, V6RouteList,
+    Session, V4RouteList, V6RouteList,
 };
 use connlib_shared::get_user_agent;
 use ip_network::{Ipv4Network, Ipv6Network};
@@ -21,6 +22,7 @@ use std::{
 use tokio::runtime::Runtime;
 use tracing_subscriber::EnvFilter;
 use tracing_subscriber::{prelude::*, util::TryInitError};
+use tun::Tun;
 
 /// The Apple client implements reconnect logic in the upper layer using OS provided
 /// APIs to detect network connectivity changes. The reconnect timeout here only
@@ -216,7 +218,7 @@ impl WrappedSession {
         )
         .map_err(|e| e.to_string())?;
         let session = Session::connect(args, portal, runtime.handle().clone());
-        session.set_tun(Tun::new().map_err(|e| e.to_string())?);
+        session.set_tun(Box::new(Tun::new().map_err(|e| e.to_string())?));
 
         Ok(Self {
             inner: session,
