+++ rust/connlib/tunnel/Cargo.toml
@@ -35,6 +35,7 @@ socket2 = { workspace = true }
 thiserror = { version = "1.0", default-features = false }
 tokio = { workspace = true }
 tracing = { workspace = true }
+tun = { workspace = true }
 
 [dev-dependencies]
 derivative = "2.2.0"
