+++ rust/connlib/tunnel/src/device_channel.rs
@@ -1,35 +1,10 @@
-#[cfg(any(target_os = "macos", target_os = "ios"))]
-mod tun_darwin;
-#[cfg(any(target_os = "macos", target_os = "ios"))]
-use tun_darwin as tun;
-
-#[cfg(target_os = "linux")]
-mod tun_linux;
-#[cfg(target_os = "linux")]
-use tun_linux as tun;
-
-#[cfg(target_os = "windows")]
-mod tun_windows;
-#[cfg(target_os = "windows")]
-use tun_windows as tun;
-
-// TODO: Android and linux are nearly identical; use a common tunnel module?
-#[cfg(target_os = "android")]
-mod tun_android;
-#[cfg(target_os = "android")]
-use tun_android as tun;
-
-#[cfg(target_family = "unix")]
-mod utils;
-
 use ip_packet::{IpPacket, MutableIpPacket, Packet as _};
 use std::io;
 use std::task::{Context, Poll, Waker};
-
-pub use tun::Tun;
+use tun::TunTrait;
 
 pub struct Device {
-    tun: Option<Tun>,
+    tun: Option<Box<dyn TunTrait>>,
     waker: Option<Waker>,
 }
 
@@ -41,7 +16,7 @@ impl Device {
         }
     }
 
-    pub(crate) fn set_tun(&mut self, tun: Tun) {
+    pub(crate) fn set_tun(&mut self, tun: Box<dyn TunTrait>) {
         tracing::info!(name = %tun.name(), "Initializing TUN device");
 
         self.tun = Some(tun);
@@ -128,45 +103,15 @@ impl Device {
         }
     }
 
-    fn tun(&self) -> io::Result<&Tun> {
-        self.tun.as_ref().ok_or_else(io_error_not_initialized)
+    fn tun(&self) -> io::Result<&dyn TunTrait> {
+        Ok(self
+            .tun
+            .as_ref()
+            .ok_or_else(io_error_not_initialized)?
+            .as_ref())
     }
 }
 
 fn io_error_not_initialized() -> io::Error {
     io::Error::new(io::ErrorKind::NotConnected, "device is not initialized yet")
-}
-
-#[cfg(any(target_os = "linux", target_os = "android"))]
-mod ioctl {
-    use super::*;
-    use std::os::fd::RawFd;
-
-    /// Executes the `ioctl` syscall on the given file descriptor with the provided request.
-    ///
-    /// # Safety
-    ///
-    /// The file descriptor must be open.
-    pub(crate) unsafe fn exec<P>(
-        fd: RawFd,
-        code: libc::c_ulong,
-        req: &mut Request<P>,
-    ) -> io::Result<()> {
-        let ret = unsafe { libc::ioctl(fd, code as _, req) };
-
-        if ret < 0 {
-            return Err(io::Error::last_os_error());
-        }
-
-        Ok(())
-    }
-
-    /// Represents a control request to an IO device, addresses by the device's name.
-    ///
-    /// The payload MUST also be `#[repr(C)]` and its layout depends on the particular request you are sending.
-    #[repr(C)]
-    pub(crate) struct Request<P> {
-        pub(crate) name: [std::ffi::c_uchar; libc::IF_NAMESIZE],
-        pub(crate) payload: P,
-    }
 }
