+++ rust/tun/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "tun"
+version = "0.1.0"
+edition = "2021"
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+libc = "0.2"
+tokio = { workspace = true }
+
+[lints]
+workspace = true
